requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
astroid==2.15.6
atomicwrites==1.4.1
attrs==22.1.0
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
Django==3.2
django-colorfield==0.9.0
django-cors-headers==3.13.0
django-debug-toolbar==3.2.4
django-extensions==3.2.3
django-filter==2.4.0
django-import-export==3.2.0
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-composed-permissions==0.2.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
Faker==12.0.1
filetype==1.2.0
flake8==3.9.2
flake8-docstrings==1.6.0
flake8-isort==6.0.0
fpdf==1.7.2
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupPy==1.14
MarkupSafe==2.1.3
mccabe==0.6.1
mixer==7.1.2
more-itertools==9.0.0
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==22.0
Pillow==9.0.0
platformdirs==3.10.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.3.1
PyJWT==2.8.0
pylint==2.17.5
pypi==2.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-environ==0.4.54
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
reportlab==3.6.11
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.0.290
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sorl-thumbnail==12.6.3
sqlparse==0.4.4
tablib==3.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.2
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.6
webcolors==1.12
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0